ai-social-media-agent/requirements.txt
streamlit
openai
pandas
numpy
requests
python-dotenv
pytest
langgraph
langchain
langchain-openai
langchain-groq
pydantic
typing-extensions